Moves getPositionOffsetTransformerFromTextModel into its own file
...to avoid core data structure files depending on ITextModel.
